pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Kirill Tuchkin <volkirtu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "3.0.0b7"
requests = "^2.29.0"
dadata = "^21.10.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
